Checkin for LivingSocial Assignment
The app “Deal Finder” does the following
1. onResume loads the data from the JSON feed in a custom list view.
2. Each row represents the deal. It shows the associated image and deal
details
3. On click of each row OR on click of the image the user is taken to
the URL associated with the deal in a browser.
4. JSON feed also sends the user information in each object. The
assumption here is that we are showing deals for the logged in user.
Hence the user information is shown in the Settings menu

What the app does not do
1. It does not cache any data locally. It fetches data every time it
resumes. Performance enhancements can be added to store the deals
locally. The thumbnails which are downloaded can also be cached.
2. Better launcher icon can be created using graphics tools.